Fix type name typo (Float32 -> Float64) (#1146)

+++ src/prompt/variable.rs
@@ -100,7 +100,7 @@ pub struct Float64;
 impl VariableInput for Float64 {
     fn type_name(&self) -> &str { "float64" }
     fn parse(&self, input: &str) -> Result<Value, Error> {
-        Ok(Value::Float32(input.parse().map_err(no_pos_err)?))
+        Ok(Value::Float64(input.parse().map_err(no_pos_err)?))
     }
 }
